requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
click==7.1.1
distlib==0.3.0
distro==1.5.0
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.3
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
tabula-py==2.1.0
virtualenv==20.0.13
Werkzeug==1.0.1
zipp==3.1.0
